refactor: TODO.bugs/29 — extract ValidateAscii inner classes to separate files
validate_ascii.rb was 549 lines with 3 sibling classes defined inline.
Extracted to:
- non_ascii_character.rb (50 lines)
- file_violations.rb (71 lines)
- non_ascii_violation_collection.rb (349 lines)

validate_ascii.rb now contains only the ValidateAscii command class
(94 lines). Autoloads added to commands.rb.

TODO 28 (Package god class) remains deferred — extracting display
methods risks breaking Thor's command dispatch pattern.

All 1454 examples pass.
