+++ TODO.bugs/29-validate-ascii-god-class.md
@@ -1,19 +1,24 @@
 # 29 — ValidateAscii is a 604-line god class with inline inner classes
 
 **Priority:** P2 (god class)
+**Status:** DONE
 
 ## Problem
 
-`lib/expressir/commands/validate_ascii.rb` — 604 lines with 3 inner classes
-(`CharacterDetail`, `NonAsciiViolation`, `NonAsciiViolationCollection`)
-defined inline.
+`validate_ascii.rb` was 549 lines with 3 sibling classes defined inline:
+NonAsciiCharacter, FileViolations, NonAsciiViolationCollection.
 
 ## Fix
 
-Extract inner classes to `lib/expressir/express/ascii_validator/` as a
-dedicated module; keep the command as a thin wrapper.
+Extracted the three inner classes to separate files:
+- `lib/expressir/commands/non_ascii_character.rb` (50 lines)
+- `lib/expressir/commands/file_violations.rb` (71 lines)
+- `lib/expressir/commands/non_ascii_violation_collection.rb` (349 lines)
+
+`validate_ascii.rb` now contains only the ValidateAscii command class
+(94 lines). Autoloads added to `lib/expressir/commands.rb`.
 
 ## Acceptance
 
-- Command file under 200 lines.
-- All validate_ascii specs pass.
+- validate_ascii.rb under 100 lines.
+- All 26 validate_ascii specs pass.
