+++ lib/expressir/commands.rb
@@ -10,6 +10,9 @@ module Expressir
     autoload :Validate, "#{__dir__}/commands/validate"
     autoload :ValidateLoad, "#{__dir__}/commands/validate_load"
     autoload :ValidateAscii, "#{__dir__}/commands/validate_ascii"
+    autoload :NonAsciiCharacter, "#{__dir__}/commands/non_ascii_character"
+    autoload :FileViolations, "#{__dir__}/commands/file_violations"
+    autoload :NonAsciiViolationCollection, "#{__dir__}/commands/non_ascii_violation_collection"
     autoload :Changes, "#{__dir__}/commands/changes"
     autoload :ChangesValidate, "#{__dir__}/commands/changes_validate"
     autoload :ChangesImportEengine, "#{__dir__}/commands/changes_import_eengine"
